+++ src/editor/nodes/EditorNodeMixin.js
@@ -12,6 +12,24 @@ import LoadingCube from "../objects/LoadingCube";
 import ErrorIcon from "../objects/ErrorIcon";
 import traverseFilteredSubtrees from "../utils/traverseFilteredSubtrees";
 
+function updateObjectArrayProperty(property, value) {
+  if (value) {
+    const compKeys = Object.keys(value);
+    compKeys.forEach(compKey => {
+      const enabledProps = new Set(property[compKey]);
+      const propKeys = Object.keys(value[compKey]);
+      propKeys.forEach(propKey => {
+        if (value[compKey][propKey] === true) {
+          enabledProps.add(propKey);
+        } else {
+          enabledProps.delete(propKey);
+        }
+      });
+      property[compKey] = Array.from(enabledProps);
+    });
+  }
+}
+
 export default function EditorNodeMixin(Object3DClass) {
   return class extends Object3DClass {
     static nodeName = "Unknown Node";
@@ -26,6 +44,8 @@ export default function EditorNodeMixin(Object3DClass) {
     static initialElementProps = {};
 
     static hideInElementsPanel = false;
+
+    static optionalProperties = {};
 
     static canAddNode(_editor) {
       return true;
@@ -68,6 +88,10 @@ export default function EditorNodeMixin(Object3DClass) {
             editorSettingsComponent.props.modifiedProperties === undefined
               ? {}
               : JSON.parse(JSON.stringify(editorSettingsComponent.props.modifiedProperties));
+          node._enabledProperties =
+            editorSettingsComponent.props.enabledProperties === undefined
+              ? {}
+              : JSON.parse(JSON.stringify(editorSettingsComponent.props.enabledProperties));
         }
       }
 
@@ -94,6 +118,7 @@ export default function EditorNodeMixin(Object3DClass) {
       this.errorIcon = null;
       this.issues = [];
       this._modifiedProperties = {};
+      this._enabledProperties = {};
     }
 
     clone(recursive) {
@@ -127,6 +152,8 @@ export default function EditorNodeMixin(Object3DClass) {
       this.enabled = source.enabled;
       this._modifiedProperties =
         source._modifiedProperties === undefined ? {} : JSON.parse(JSON.stringify(source._modifiedProperties));
+      this._enabledProperties =
+        source._enabledProperties === undefined ? {} : JSON.parse(JSON.stringify(source._enabledProperties));
 
       return this;
     }
@@ -196,7 +223,9 @@ export default function EditorNodeMixin(Object3DClass) {
             props: {
               enabled: this.enabled,
               modifiedProperties:
-                this._modifiedProperties === undefined ? {} : JSON.parse(JSON.stringify(this._modifiedProperties))
+                this._modifiedProperties === undefined ? {} : JSON.parse(JSON.stringify(this._modifiedProperties)),
+              enabledProperties:
+                this._enabledProperties === undefined ? {} : JSON.parse(JSON.stringify(this._enabledProperties))
             }
           }
         ]
@@ -467,33 +496,46 @@ export default function EditorNodeMixin(Object3DClass) {
       }
     }
 
-    optionalPropertyExportValue(componentName, propName) {
+    enabledPropertyExportValue(componentName, propName) {
+      if (this._enabledProperties[componentName]) {
+        return this._enabledProperties[componentName].includes(propName) ? this[propName] : undefined;
+      }
+      return undefined;
+    }
+
+    modifiedPropertyExportValue(componentName, propName) {
       if (this._modifiedProperties[componentName]) {
         return this._modifiedProperties[componentName].includes(propName) ? this[propName] : undefined;
       }
       return undefined;
     }
 
-    set modifiedProperties(object) {
-      if (object) {
-        const compKeys = Object.keys(object);
-        compKeys.forEach(compKey => {
-          const modifiedProps = new Set(this._modifiedProperties[compKey]);
-          const propKeys = Object.keys(object[compKey]);
-          propKeys.forEach(propKey => {
-            if (object[compKey][propKey] === true) {
-              modifiedProps.add(propKey);
-            } else {
-              modifiedProps.delete(propKey);
-            }
-          });
-          this._modifiedProperties[compKey] = Array.from(modifiedProps);
-        });
+    exportPropertyValue(componentName, propName) {
+      if (this.constructor.optionalProperties[componentName]) {
+        if (this.constructor.optionalProperties[componentName].includes(propName)) {
+          return this.enabledPropertyExportValue(componentName, propName);
+        } else {
+          return this.modifiedPropertyExportValue(componentName, propName);
+        }
+      } else {
+        return this.modifiedPropertyExportValue(componentName, propName);
       }
+    }
+
+    set modifiedProperties(object) {
+      updateObjectArrayProperty(this._modifiedProperties, object);
     }
 
     get modifiedProperties() {
       return this._modifiedProperties;
+    }
+
+    set enabledProperties(object) {
+      updateObjectArrayProperty(this._enabledProperties, object);
+    }
+
+    get enabledProperties() {
+      return this._enabledProperties;
     }
   };
 }
