+++ src/editor/nodes/SceneNode.js
@@ -8,391 +8,15 @@ import getNodeWithUUID from "../utils/getNodeWithUUID";
 import serializeColor from "../utils/serializeColor";
 import { AvatarAudioDefaults, MediaAudioDefaults } from "../objects/AudioParams";
 import traverseFilteredSubtrees from "../utils/traverseFilteredSubtrees";
-
-// Migrate v1 spoke scene to v2
-function migrateV1ToV2(json) {
-  const { root, metadata, entities } = json;
-
-  // Generate UUIDs for all existing entity names.
-  const rootUUID = _Math.generateUUID();
-  const nameToUUID = { [root]: rootUUID };
-  for (const name in entities) {
-    if (Object.prototype.hasOwnProperty.call(entities, name)) {
-      nameToUUID[name] = _Math.generateUUID();
-    }
-  }
-
-  // Replace names with uuids in entities and add the name property.
-  const newEntities = { [rootUUID]: { name: root } };
-  for (const [name, entity] of Object.entries(entities)) {
-    const uuid = nameToUUID[name];
-    newEntities[uuid] = Object.assign({}, entity, {
-      name,
-      parent: nameToUUID[entity.parent]
-    });
-  }
-
-  return {
-    version: 2,
-    root: nameToUUID[root],
-    entities: newEntities,
-    metadata
-  };
-}
-
-function migrateV2ToV3(json) {
-  json.version = 3;
-
-  for (const entityId in json.entities) {
-    if (!Object.prototype.hasOwnProperty.call(json.entities, entityId)) continue;
-
-    const entity = json.entities[entityId];
-
-    if (!entity.components) {
-      continue;
-    }
-
-    entity.components.push({
-      name: "visible",
-      props: {
-        value: true
-      }
-    });
-
-    const modelComponent = entity.components.find(c => c.name === "gltf-model");
-    const navMeshComponent = entity.components.find(c => c.name === "nav-mesh");
-
-    if (!navMeshComponent && modelComponent && modelComponent.props.includeInFloorPlan) {
-      entity.components.push({
-        name: "collidable",
-        props: {}
-      });
-
-      entity.components.push({
-        name: "walkable",
-        props: {}
-      });
-    }
-
-    const groundPlaneComponent = entity.components.find(c => c.name === "ground-plane");
-
-    if (groundPlaneComponent) {
-      entity.components.push({
-        name: "walkable",
-        props: {}
-      });
-    }
-
-    if (modelComponent && navMeshComponent) {
-      entity.components = [
-        {
-          name: "floor-plan",
-          props: {
-            autoCellSize: true,
-            cellSize: 1,
-            cellHeight: 0.1,
-            agentHeight: 1.0,
-            agentRadius: 0.0001,
-            agentMaxClimb: 0.5,
-            agentMaxSlope: 45,
-            regionMinSize: 4
-          }
-        }
-      ];
-    }
-  }
-
-  return json;
-}
-
-function migrateV3ToV4(json) {
-  json.version = 4;
-
-  for (const entityId in json.entities) {
-    if (!Object.prototype.hasOwnProperty.call(json.entities, entityId)) continue;
-
-    const entity = json.entities[entityId];
-
-    if (!entity.components) {
-      continue;
-    }
-
-    const visibleComponent = entity.components.find(c => c.name === "visible");
-
-    if (visibleComponent) {
-      if (visibleComponent.props.visible !== undefined) {
-        continue;
-      }
-
-      if (visibleComponent.props.value !== undefined) {
-        visibleComponent.props = {
-          visible: visibleComponent.props.value
-        };
-      } else {
-        visibleComponent.props = {
-          visible: true
-        };
-      }
-    }
-  }
-
-  return json;
-}
-
-const combineComponents = ["gltf-model", "kit-piece"];
-
-function migrateV4ToV5(json) {
-  json.version = 5;
-
-  for (const entityId in json.entities) {
-    if (!Object.prototype.hasOwnProperty.call(json.entities, entityId)) continue;
-
-    const entity = json.entities[entityId];
-
-    if (!entity.components) {
-      continue;
-    }
-
-    const animationComponent = entity.components.find(c => c.name === "loop-animation");
-
-    if (animationComponent) {
-      // Prior to V5 animation clips were stored in activeClipIndex as an integer
-      const { activeClipIndex } = animationComponent.props;
-      delete animationComponent.props.activeClipIndex;
-      // In V5+ activeClipIndices stores an array of integers. It may be undefined if migrating from a legacy scene where the
-      // clip property stores the animation clip name. We can't migrate this here so we do it in ModelNode and KitPieceNode.
-      animationComponent.props.activeClipIndices = activeClipIndex !== undefined ? [activeClipIndex] : [];
-    }
-
-    const hasCombineComponent = entity.components.find(c => combineComponents.indexOf(c.name) !== -1);
-
-    if (hasCombineComponent) {
-      entity.components.push({
-        name: "combine",
-        props: {}
-      });
-    }
-  }
-
-  return json;
-}
-
-function migrateV5ToV6(json) {
-  json.version = 6;
-
-  for (const entityId in json.entities) {
-    if (!Object.prototype.hasOwnProperty.call(json.entities, entityId)) continue;
-
-    const entity = json.entities[entityId];
-
-    if (!entity.components) {
-      continue;
-    }
-
-    const audioComponent = entity.components.find(c => c.name === "audio");
-
-    if (audioComponent) {
-      // Prior to V6 audio parameters where part of the audio node, now they have been refactored to the audio-params component
-      entity.components.push({
-        name: "audio-params",
-        props: {
-          audioType: audioComponent.props["audioType"],
-          gain: audioComponent.props["volume"],
-          distanceModel: audioComponent.props["distanceModel"],
-          rolloffFactor: audioComponent.props["rolloffFactor"],
-          refDistance: audioComponent.props["refDistance"],
-          maxDistance: audioComponent.props["maxDistance"],
-          coneInnerAngle: audioComponent.props["coneInnerAngle"],
-          coneOuterAngle: audioComponent.props["coneOuterAngle"],
-          coneOuterGain: audioComponent.props["coneOuterGain"]
-        }
-      });
-
-      delete audioComponent.props["audioType"];
-      delete audioComponent.props["volume"];
-      delete audioComponent.props["distanceModel"];
-      delete audioComponent.props["rolloffFactor"];
-      delete audioComponent.props["refDistance"];
-      delete audioComponent.props["maxDistance"];
-      delete audioComponent.props["coneInnerAngle"];
-      delete audioComponent.props["coneOuterAngle"];
-      delete audioComponent.props["coneOuterGain"];
-    }
-
-    const videoComponent = entity.components.find(c => c.name === "video");
-
-    if (videoComponent) {
-      // Prior to V6 audio parameters where part of the audio node, now they have been refactored to the audio-params component
-      entity.components.push({
-        name: "audio-params",
-        props: {
-          audioType: videoComponent.props["audioType"],
-          gain: videoComponent.props["volume"],
-          distanceModel: videoComponent.props["distanceModel"],
-          rolloffFactor: videoComponent.props["rolloffFactor"],
-          refDistance: videoComponent.props["refDistance"],
-          maxDistance: videoComponent.props["maxDistance"],
-          coneInnerAngle: videoComponent.props["coneInnerAngle"],
-          coneOuterAngle: videoComponent.props["coneOuterAngle"],
-          coneOuterGain: videoComponent.props["coneOuterGain"]
-        }
-      });
-
-      delete videoComponent.props["audioType"];
-      delete videoComponent.props["gain"];
-      delete videoComponent.props["distanceModel"];
-      delete videoComponent.props["rolloffFactor"];
-      delete videoComponent.props["refDistance"];
-      delete videoComponent.props["maxDistance"];
-      delete videoComponent.props["coneInnerAngle"];
-      delete videoComponent.props["coneOuterAngle"];
-      delete videoComponent.props["coneOuterGain"];
-    }
-  }
-
-  return json;
-}
-
-function migrateV6ToV7(json) {
-  json.version = 7;
-
-  for (const entityId in json.entities) {
-    if (!Object.prototype.hasOwnProperty.call(json.entities, entityId)) continue;
-
-    const entity = json.entities[entityId];
-
-    if (!entity.components) {
-      continue;
-    }
-
-    const audioParamsComponent = entity.components.find(c => c.name === "audio-params");
-    if (audioParamsComponent) {
-      // Prior to V6 we didn't have dirty params so we need to enable all properties
-      // to make sure that the old settings are applied.
-      let editorSettingsComponent = entity.components.find(c => c.name === "editor-settings");
-      if (!editorSettingsComponent) {
-        editorSettingsComponent = {
-          name: "editor-settings",
-          props: {
-            enabled: true
-          }
-        };
-        entity.components.push(editorSettingsComponent);
-      }
-      editorSettingsComponent.props["modifiedProperties"] = {
-        "audio-params": {
-          audioType: true,
-          gain: true,
-          distanceModel: true,
-          rolloffFactor: true,
-          refDistance: true,
-          maxDistance: true,
-          coneInnerAngle: true,
-          coneOuterAngle: true,
-          coneOuterGain: true
-        }
-      };
-    }
-
-    const audioSettingsComponent = entity.components.find(c => c.name === "audio-settings");
-    if (audioSettingsComponent) {
-      const overriden = audioSettingsComponent.props && audioSettingsComponent.props.overrideAudioSettings;
-      // Prior to V6 we didn't have dirty params so we need to enable all properties
-      // to make sure that the old settings are applied.
-      if (overriden) {
-        let editorSettingsComponent = entity.components.find(c => c.name === "editor-settings");
-        if (!editorSettingsComponent) {
-          editorSettingsComponent = {
-            name: "editor-settings",
-            props: {
-              enabled: true
-            }
-          };
-          entity.components.push(editorSettingsComponent);
-        }
-        editorSettingsComponent.props["modifiedProperties"] = {
-          scene: {
-            avatarDistanceModel: true,
-            avatarRolloffFactor: true,
-            avatarRefDistance: true,
-            avatarMaxDistance: true,
-            mediaVolume: true,
-            mediaDistanceModel: true,
-            mediaRolloffFactor: true,
-            mediaRefDistance: true,
-            mediaMaxDistance: true,
-            mediaConeInnerAngle: true,
-            mediaConeOuterAngle: true,
-            mediaConeOuterGain: true
-          }
-        };
-      }
-    }
-  }
-
-  return json;
-}
-
-function migrateV7ToV8(json) {
-  json.version = 8;
-
-  for (const entityId in json.entities) {
-    if (!Object.prototype.hasOwnProperty.call(json.entities, entityId)) continue;
-
-    const entity = json.entities[entityId];
-
-    if (!entity.components) {
-      continue;
-    }
-
-    const updateModifiedProps = oldProps => {
-      const compKeys = Object.keys(oldProps["modifiedProperties"]);
-      compKeys.forEach(compKey => {
-        const newModifiedProps = [];
-        const propKeys = Object.keys(oldProps["modifiedProperties"][compKey]);
-        propKeys.forEach(propKey => {
-          if (oldProps["modifiedProperties"][compKey][propKey] === true) {
-            newModifiedProps.push(propKey);
-          }
-        });
-        oldProps["modifiedProperties"][compKey] = newModifiedProps;
-      });
-    };
-
-    const audioParamsComponent = entity.components.find(c => c.name === "audio-params");
-    if (audioParamsComponent) {
-      // Migrate old modified params object to a set
-      const editorSettingsComponent = entity.components.find(c => c.name === "editor-settings");
-      const modifiedProps = editorSettingsComponent.props["modifiedProperties"];
-      if (modifiedProps) {
-        updateModifiedProps(editorSettingsComponent.props);
-      }
-    }
-
-    const audioSettingsComponent = entity.components.find(c => c.name === "audio-settings");
-    if (audioSettingsComponent) {
-      const overriden = audioSettingsComponent.props && audioSettingsComponent.props.overrideAudioSettings;
-      // Migrate old modified params object to a se
-      if (overriden) {
-        const editorSettingsComponent = entity.components.find(c => c.name === "editor-settings");
-        if (editorSettingsComponent) {
-          const modifiedProps = editorSettingsComponent.props["modifiedProperties"];
-          if (modifiedProps) {
-            updateModifiedProps(editorSettingsComponent.props);
-          }
-        }
-      }
-    }
-  }
-
-  return json;
-}
+import MigrateScene from "./SceneMigration";
 
 export const FogType = {
   Disabled: "disabled",
   Linear: "linear",
   Exponential: "exponential"
 };
+
+const JSON_VERSION = 9;
 
 export default class SceneNode extends EditorNodeMixin(Scene) {
   static nodeName = "Scene";
@@ -404,33 +28,7 @@ export default class SceneNode extends EditorNodeMixin(Scene) {
   }
 
   static async loadProject(editor, json) {
-    if (!json.version) {
-      json = migrateV1ToV2(json);
-    }
-
-    if (json.version === 2) {
-      json = migrateV2ToV3(json);
-    }
-
-    if (json.version === 3) {
-      json = migrateV3ToV4(json);
-    }
-
-    if (json.version === 4) {
-      json = migrateV4ToV5(json);
-    }
-
-    if (json.version === 5) {
-      json = migrateV5ToV6(json);
-    }
-
-    if (json.version === 6) {
-      json = migrateV6ToV7(json);
-    }
-
-    if (json.version === 7) {
-      json = migrateV7ToV8(json);
-    }
+    json = MigrateScene(json, JSON_VERSION);
 
     const { root, metadata, entities } = json;
 
@@ -669,7 +267,7 @@ export default class SceneNode extends EditorNodeMixin(Scene) {
 
   serialize() {
     const sceneJson = {
-      version: 8,
+      version: JSON_VERSION,
       root: this.uuid,
       metadata: JSON.parse(JSON.stringify(this.metadata)),
       entities: {
@@ -786,18 +384,18 @@ export default class SceneNode extends EditorNodeMixin(Scene) {
     }
 
     if (this.overrideAudioSettings) {
-      const avatarDistanceModel = this.optionalPropertyExportValue("scene", "avatarDistanceModel");
-      const avatarRolloffFactor = this.optionalPropertyExportValue("scene", "avatarRolloffFactor");
-      const avatarRefDistance = this.optionalPropertyExportValue("scene", "avatarRefDistance");
-      const avatarMaxDistance = this.optionalPropertyExportValue("scene", "avatarMaxDistance");
-      const mediaVolume = this.optionalPropertyExportValue("scene", "mediaVolume");
-      const mediaDistanceModel = this.optionalPropertyExportValue("scene", "mediaDistanceModel");
-      const mediaRolloffFactor = this.optionalPropertyExportValue("scene", "mediaRolloffFactor");
-      const mediaRefDistance = this.optionalPropertyExportValue("scene", "mediaRefDistance");
-      const mediaMaxDistance = this.optionalPropertyExportValue("scene", "mediaMaxDistance");
-      const mediaConeInnerAngle = this.optionalPropertyExportValue("scene", "mediaConeInnerAngle");
-      const mediaConeOuterAngle = this.optionalPropertyExportValue("scene", "mediaConeOuterAngle");
-      const mediaConeOuterGain = this.optionalPropertyExportValue("scene", "mediaConeOuterGain");
+      const avatarDistanceModel = this.exportPropertyValue("scene", "avatarDistanceModel");
+      const avatarRolloffFactor = this.exportPropertyValue("scene", "avatarRolloffFactor");
+      const avatarRefDistance = this.exportPropertyValue("scene", "avatarRefDistance");
+      const avatarMaxDistance = this.exportPropertyValue("scene", "avatarMaxDistance");
+      const mediaVolume = this.exportPropertyValue("scene", "mediaVolume");
+      const mediaDistanceModel = this.exportPropertyValue("scene", "mediaDistanceModel");
+      const mediaRolloffFactor = this.exportPropertyValue("scene", "mediaRolloffFactor");
+      const mediaRefDistance = this.exportPropertyValue("scene", "mediaRefDistance");
+      const mediaMaxDistance = this.exportPropertyValue("scene", "mediaMaxDistance");
+      const mediaConeInnerAngle = this.exportPropertyValue("scene", "mediaConeInnerAngle");
+      const mediaConeOuterAngle = this.exportPropertyValue("scene", "mediaConeOuterAngle");
+      const mediaConeOuterGain = this.exportPropertyValue("scene", "mediaConeOuterGain");
       this.addGLTFComponent("audio-settings", {
         ...(avatarDistanceModel && { avatarDistanceModel }),
         ...(avatarRolloffFactor && { avatarRolloffFactor }),
