+++ src/ui/inputs/InputGroup.js
@@ -4,6 +4,7 @@ import styled from "styled-components";
 import { QuestionCircle } from "styled-icons/fa-regular/QuestionCircle";
 import { InfoTooltip } from "../layout/Tooltip";
 import ResetButton from "./ResetButton";
+import BooleanInput from "./BooleanInput";
 import { PropertyLabel } from "./PropertyLabel";
 
 export const InputGroupContainer = styled.div`
@@ -12,6 +13,7 @@ export const InputGroupContainer = styled.div`
   padding: 4px 8px;
   flex: 1;
   min-height: 24px;
+  align-items: center;
 
   ${props =>
     props.disabled &&
@@ -22,7 +24,6 @@ export const InputGroupContainer = styled.div`
 
   & > label {
     display: block;
-    width: 25%;
     color: ${props => props.theme.text2};
     padding-bottom: 2px;
     padding-top: 4px;
@@ -30,9 +31,15 @@ export const InputGroupContainer = styled.div`
 `;
 
 export const InputGroupContent = styled.div`
+  ${props =>
+    props.disabled &&
+    `
+    pointer-events: none;
+    opacity: 0.3;
+  `}
   display: flex;
   flex-direction: row;
-  flex: 1;
+  flex: 2;
   padding-left: 8px;
   align-items: center;
 `;
@@ -44,6 +51,38 @@ export const InputGroupInfoIcon = styled(QuestionCircle)`
   color: ${props => props.theme.blue};
   cursor: pointer;
   align-self: center;
+`;
+
+export const InputGroupHeader = styled.div`
+  display: flex;
+  flex-direction: row;
+  flex: 1;
+  align-items: center;
+
+  ${props =>
+    props.disabled &&
+    `
+    pointer-events: none;
+    opacity: 0.3;
+  `}
+
+  & > :first-child {
+    padding-right: 8px;
+  }
+`;
+
+export const OptionalGroup = styled.div`
+  display: flex;
+  flex-direction: row;
+  flex: 1;
+  align-items: center;
+
+  ${props =>
+    props.disabled &&
+    `
+    pointer-events: none;
+    opacity: 0.3;
+  `}
 `;
 
 export function InputGroupInfo({ info }) {
@@ -58,11 +97,16 @@ InputGroupInfo.propTypes = {
   info: PropTypes.string
 };
 
-export default function InputGroup({ name, children, disabled, info, reset, onReset }) {
+export default function InputGroup({ name, children, disabled, info, optional, enabled, onEnable, reset, onReset }) {
   return (
     <InputGroupContainer disabled={disabled}>
-      <PropertyLabel modified={!reset}>{name}:</PropertyLabel>
-      <InputGroupContent>
+      <InputGroupHeader>
+        {optional && <BooleanInput value={enabled} onChange={onEnable} />}
+        <OptionalGroup disabled={optional && !enabled}>
+          {name && <PropertyLabel modified={!reset}>{name}:</PropertyLabel>}
+        </OptionalGroup>
+      </InputGroupHeader>
+      <InputGroupContent disabled={optional && !enabled}>
         {children}
         {info && <InputGroupInfo info={info} />}
         {onReset && <ResetButton disabled={!reset} onClick={onReset} />}
@@ -77,6 +121,9 @@ InputGroup.propTypes = {
   disabled: PropTypes.bool,
   className: PropTypes.string,
   info: PropTypes.string,
+  optional: PropTypes.bool,
+  enabled: PropTypes.bool,
+  onEnable: PropTypes.func,
   onReset: PropTypes.func,
   reset: PropTypes.bool
 };
